Add SBOM JSON files for ATfE, llvmlibc overlay and newlib overlay; and Python scripts which generate these SBOM JSON files.
