--- arm-software/embedded/scripts/SBOM_Files/spdx2_document_from_scratch_ATfE.py
@@ -1,300 +1,0 @@
-#  SPDX-FileCopyrightText: 2023 spdx contributors
-#
-#  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-#
-# Purpose: This script helps to generate SBOM for ATfE.
-#
-# Installation: Kindly refer https://github.com/spdx/tools-python for pre-requisite installation.
-#
-# Usage: python spdx2_document_from_scratch_ATfE.py
-#
-# Output file generated: sbom_atfe.spdx.json
-#
-import logging
-import uuid
-from datetime import datetime
-from typing import List
-
-from spdx_tools.common.spdx_licensing import spdx_licensing
-from spdx_tools.spdx.model import (
-    Actor,
-    ActorType,
-    CreationInfo,
-    Document,
-    Package,
-    Relationship,
-    RelationshipType,
-)
-from spdx_tools.spdx.validation.document_validator import validate_full_spdx_document
-from spdx_tools.spdx.validation.validation_message import ValidationMessage
-from spdx_tools.spdx.writer.write_anything import write_file
-
-# This example shows how to use the spdx-tools to create an SPDX document from scratch,
-# validate it and write it to a file.
-
-def create_Package(
-    name,
-    uuid,
-    download_location,
-    actorType,
-    personName,
-    personEmail,
-    license,
-    licenseComment,
-):
-    package = Package(
-        name=name,
-        spdx_id=uuid,
-        download_location=download_location,
-        supplier=Actor(actorType, personName, personEmail),
-        originator=Actor(actorType, personName, personEmail),
-        files_analyzed=False,
-        license_concluded=spdx_licensing.parse(license),
-        license_declared=spdx_licensing.parse(license),
-        license_comment=licenseComment,
-    )
-
-    return package
-
-# First up, we need general information about the creation of the document, summarised by the CreationInfo class.
-creation_info = CreationInfo(
-    spdx_version="SPDX-2.3",
-    spdx_id="SPDXRef-DOCUMENT",
-    name="Arm Toolchain for Embedded",
-    data_license="CC0-1.0",
-    document_namespace="https://github.com/arm/arm-toolchain/tree/arm-software/arm-software/embedded",
-    creators=[
-        Actor(ActorType.ORGANIZATION, "Arm Limited", "open-source-office@arm.com")
-    ],
-    created=datetime(2025, 3, 5),
-)
-
-# creation_info is the only required property of the Document class (have a look there!), the rest are optional lists.
-# So, we are set up to create a new document instance.
-document = Document(creation_info)
-
-# The document currently does not describe anything. Let's create a package that we can add to it.
-# The Package class has quite a few properties (have a look there!),
-# but only name, spdx_id and download_location are mandatory in SPDX v2.3.
-package_url = (
-    "https://github.com/arm/arm-toolchain/tree/arm-software/arm-software/embedded"
-)
-package_uuid_atfe = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_atfe = create_Package(
-    "Arm Toolchain for Embedded",
-    package_uuid_atfe,
-    package_url,
-    ActorType.ORGANIZATION,
-    "Arm Limited",
-    "open-source-office@arm.com",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project"
-package_uuid_llvm = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, "https://github.com/llvm/llvm-project")
-)
-package_llvm = create_Package(
-    "llvm-project",
-    package_uuid_llvm,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/clang"
-package_uuid_clang = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_clang = create_Package(
-    "clang",
-    package_uuid_clang,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/lld"
-package_uuid_lld = "SPDXRef-Package-" + str(uuid.uuid5(uuid.NAMESPACE_URL, package_url))
-package_lld = create_Package(
-    "lld",
-    package_uuid_lld,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/compiler-rt"
-package_uuid_compiler_rt = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_compiler_rt = create_Package(
-    "compiler-rt",
-    package_uuid_compiler_rt,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/libcxx"
-package_uuid_libcxx = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_libcxx = create_Package(
-    "libcxx",
-    package_uuid_libcxx,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/libcxxabi"
-package_uuid_libcxxabi = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_libcxxabi = create_Package(
-    "libcxxabi",
-    package_uuid_libcxxabi,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/llvm/llvm-project/tree/main/libunwind"
-package_uuid_libunwind = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_libunwind = create_Package(
-    "libunwind",
-    package_uuid_libunwind,
-    package_url,
-    ActorType.ORGANIZATION,
-    "https://foundation.llvm.org/",
-    "info@llvm.org",
-    "Apache-2.0 WITH LLVM-exception",
-    "",
-)
-
-package_url = "https://github.com/picolibc/picolibc"
-package_uuid_picolibc = "SPDXRef-Package-" + str(
-    uuid.uuid5(uuid.NAMESPACE_URL, package_url)
-)
-package_picolibc = create_Package(
-    "picolibc",
-    package_uuid_picolibc,
-    package_url,
-    ActorType.PERSON,
-    "Keith Packard",
-    "https://github.com/keith-packard",
-    "",
-    "Detailed list of licenses is at: https://github.com/picolibc/picolibc/blob/main/COPYING.picolibc",
-)
-
-# Now that we have a package defined, we can add it to the document's package property.
-document.packages = [
-    package_atfe,
-    package_llvm,
-    package_clang,
-    package_lld,
-    package_compiler_rt,
-    package_libcxx,
-    package_libcxxabi,
-    package_libunwind,
-    package_picolibc,
-]
-
-# A DESCRIBES relationship asserts that the document indeed describes the package.
-describes_relationship = Relationship(
-    "SPDXRef-DOCUMENT", RelationshipType.DESCRIBES, package_uuid_atfe
-)
-document.relationships = [describes_relationship]
-
-# Assuming the package contains those two files, we create two CONTAINS relationships.
-
-# This library uses run-time type checks when assigning properties.
-# Because in-place alterations like .append() circumvent these checks, we don't use them here.
-
-# We now have created a document with basic creation information, describing a package that contains two files.
-# You can also add Annotations, Snippets and ExtractedLicensingInfo to the document in an analogous manner to the above.
-# Have a look at their respective classes if you are unsure about their properties.
-
-# Assuming the package contains those two files, we create two CONTAINS relationships.
-contains_relationship_atfe = Relationship(
-    package_uuid_atfe, RelationshipType.CONTAINS, package_uuid_atfe
-)
-contains_relationship_llvm = Relationship(
-    package_uuid_llvm, RelationshipType.CONTAINS, package_uuid_llvm
-)
-contains_relationship_clang = Relationship(
-    package_uuid_clang, RelationshipType.CONTAINS, package_uuid_clang
-)
-contains_relationship_lld = Relationship(
-    package_uuid_lld, RelationshipType.CONTAINS, package_uuid_lld
-)
-contains_relationship_compiler_rt = Relationship(
-    package_uuid_compiler_rt, RelationshipType.CONTAINS, package_uuid_compiler_rt
-)
-contains_relationship_libcxx = Relationship(
-    package_uuid_libcxx, RelationshipType.CONTAINS, package_uuid_libcxx
-)
-contains_relationship_libcxxabi = Relationship(
-    package_uuid_libcxxabi, RelationshipType.CONTAINS, package_uuid_libcxxabi
-)
-contains_relationship_libunwind = Relationship(
-    package_uuid_libunwind, RelationshipType.CONTAINS, package_uuid_libunwind
-)
-contains_relationship_picolibc = Relationship(
-    package_uuid_picolibc, RelationshipType.CONTAINS, package_uuid_picolibc
-)
-
-# Because in-place alterations like .append() circumvent these checks, we don't use them here.
-document.relationships += [
-    contains_relationship_atfe,
-    contains_relationship_llvm,
-    contains_relationship_clang,
-    contains_relationship_lld,
-    contains_relationship_compiler_rt,
-    contains_relationship_libcxx,
-    contains_relationship_libcxxabi,
-    contains_relationship_libunwind,
-    contains_relationship_picolibc,
-]
-
-# This library provides comprehensive validation against the SPDX specification.
-# Note that details of the validation depend on the SPDX version of the document.
-validation_messages: List[ValidationMessage] = validate_full_spdx_document(document)
-
-# You can have a look at each entry's message and context (like spdx_id, parent_id, full_element)
-# which will help you pinpoint the location of the invalidity.
-for message in validation_messages:
-    logging.warning(message.validation_message)
-    logging.warning(message.context)
-
-# If the document is valid, validation_messages will be empty.
-assert validation_messages == []
-
-# Finally, we can serialize the document to any of the five supported formats.
-# Using the write_file() method from the write_anything module,
-# the format will be determined by the file ending: .spdx (tag-value), .json, .xml, .yaml. or .rdf (or .rdf.xml)
-write_file(document, "sbom_atfe.spdx.json")
